requirements/common.txt
Django==1.9.1
django-debug-toolbar==1.4
django-crispy-forms==1.6.0
django-registration-redux==1.3
django-grappelli==2.7.3
django-hijack==2.0.3
